zST-model-integration-aml/requirements.txt
numba==0.57.1
shap==0.42.1
wheel==0.38.1
pybind11==2.10.4
setuptools==67.8.0
six==1.16.0
threadpoolctl==3.1.0
cython==0.29.33
meson==1.0.0
pkgconfig==1.5.5
ninja==1.11.1
pythran==0.12.1
snapml==1.14.0
xgboost==1.7.5
fastapi==0.95.2
uvicorn[standard]
requests==2.31.0
python-dotenv==1.0.0